[endpoint] Add connection to sqlite3 database
Introducing a simple sqlite3 database. This also adds some error
handling using anyhow as well. The current approach is pretty
hacky and we should find a cleaner way to do this such that we can use
`?` and implement directly within the defined function.

Simple test was done to ensure we can connect to the database. Note
the database will be created if none exists.

The DATABASE_URL environment variable needs to be added (we use dot_env
here by defining a .env file)