[server] Add the first endpoint to mark a habit as completed
This diff introduces the endpoint in routes.rs to mark a habit. This API
also creates the habit if it doesn't exist. Additionally, this diff
introduces database.rs which is a file that contains all of our SQL
queries to the database.

We also have some basic unit tests (which are really integration tests)
to verify some simple logic.