[server] Add a bad commit with lots of misc changes
Unfortunately, we hacked on this whilst developing the app. This
is less than ideal code. Since we're the only one working on the project
we determine this is the fastest way to go forward. Rather than cleaning
this up perfectly we be pragmatic and keep working on the e2e on the app
side then come back to this.

I have a feeling this is going to bite me later...