[app][account] Implement logout button
This will clear the local storage and go back to login. Note: using
the back button doesn't clear local storage.

This is somewhat a feature/bug. I assume this back button will soon
disappear in the future as a replacement for a separate button. For now
we just leave it since it still works :)