[app][flutter] Change to using nested scaffold to remove hard coding of size
Previously, we just added an additional item to our Column at the end
with a specific size to specify the current task. This was kind of hacky
since the list of tasks doesn't have to do with the additional item.

Instead we use a scaffold to allow us to have a bottom nav bar which
will always be at the bottom. The added benefit of this is snackbars
will go above this bar.

Finally, We added a FAB and fixed some colors.